requirements.txt
fastapi==0.95.2
numpy==1.23.4
Pillow==9.5.0
Pillow==9.5.0
pydantic==1.10.4
